FEATURE: Initial commit, add property-silhuettes for Neos nodeTypes
The silhuettes are configured in the Settings.path `Sitegeist.Silhuettes.properties` and
are adressed via `options.silhuette = silhuette.path` in the NodeType-property configuration.

The settings from the configured silhuette are merged with the configuration that is found in the nodeType with the
local configuration taking precedence over the silhuette.

The package brings three default silhuettes:

- `text.plain`: An inline editable string where no formatting is allowed.
- `text.block`: An inline editable string where only inline formatting is enabled.
- `text.free`: An inline editable string all formatting including blocks is allowed.